multiagenttravelsystem/core_llm_chain.py
from langchain_openai import ChatOpenAI
from langchain_core.messages import HumanMessage, AIMessage
from langchain_core.output_parsers import StrOutputParser
from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder
from langgraph.graph import StateGraph, END

llm = ChatOpenAI(model_name="gpt-4o", temperature=0.2)